--- stock_ext_sst/views/stock_picking_views.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-
-    <record id="stock_vpicktree" model="ir.ui.view">
-        <field name="name">stock.vpicktree</field>
-        <field name="model">stock.picking</field>
-        <field name="inherit_id" ref="stock.vpicktree" />
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='scheduled_date']" position="before">
-                <field name="product_id" />
-                <field name="carrier_id" />
-            </xpath>
-        </field>
-    </record>
-
-    <record id="view_picking_internal_search" model="ir.ui.view">
-        <field name="name">stock.picking.internal.search</field>
-        <field name="model">stock.picking</field>
-        <field name="inherit_id" ref="stock.view_picking_internal_search" />
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='product_id']" position="replace" />
-            <xpath expr="//field[@name='name']" position="before">
-                <field name="product_id" />
-            </xpath>
-        </field>
-    </record>
-
-</odoo>
