+++ stock_quant_attribute/__init__.py
@@ -1,2 +1,1 @@
 from . import models
-from . import wizard
