+++ stock_quant_attribute/__manifest__.py
@@ -1,7 +1,7 @@
 # Copyright 2017-2018 Quartile Limited
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 {
-    "name": "Extension for inventory functions",
+    "name": "Stock Quant Attribute",
     "version": "11.0.2.0.0",
     "author": "Quartile Limited",
     "website": "https://www.quartile.co",
@@ -10,15 +10,10 @@
     "depends": [
         "stock",
         "delivery",
-        "product_yahoo_auction_sst",
         "stock_quant_list_view",
     ],
     "data": [
-        "views/res_config_settings_views.xml",
-        "views/stock_picking_views.xml",
         "views/stock_quant_views.xml",
-        "wizard/stock_quant_publish_wizard.xml",
-        "wizard/stock_quant_status_update_wizard.xml",
     ],
     "installable": True,
 }
