+++ stock_quant_attribute/i18n/ja.po
@@ -1,0 +1,41 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+#	* stock_quant_attribute
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 11.0\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2023-10-02 12:37+0000\n"
+"PO-Revision-Date: 2023-10-02 12:37+0000\n"
+"Last-Translator: <>\n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: \n"
+
+#. module: stock_quant_attribute
+#: model:ir.model.fields,field_description:stock_quant_attribute.field_stock_quant_default_code
+msgid "Barcode"
+msgstr "Barcode"
+
+#. module: stock_quant_attribute
+#: model:ir.model.fields,help:stock_quant_attribute.field_stock_quant_list_price
+msgid "Base price to compute the customer price. Sometimes called the catalog price."
+msgstr "Base price to compute the customer price. Sometimes called the catalog price."
+
+#. module: stock_quant_attribute
+#: model:ir.model,name:stock_quant_attribute.model_stock_quant
+msgid "Quants"
+msgstr "Quants"
+
+#. module: stock_quant_attribute
+#: model:ir.model.fields,field_description:stock_quant_attribute.field_stock_quant_list_price
+msgid "Sale Price"
+msgstr "Sale Price"
+
+#. module: stock_quant_attribute
+#: model:ir.model.fields,field_description:stock_quant_attribute.field_stock_quant_website_published
+msgid "Visible in Website"
+msgstr "Visible in Website"
