+++ stock_quant_attribute/models/__init__.py
@@ -1,0 +1,1 @@
+from . import stock_quant
