+++ stock_quant_attribute/views/stock_quant_views.xml
@@ -1,19 +1,16 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <odoo>
-
     <record id="view_stock_quant_tree" model="ir.ui.view">
         <field name="name">stock.quant.tree</field>
         <field name="model">stock.quant</field>
         <field name="inherit_id" ref="stock_quant_list_view.view_stock_quant_tree" />
         <field name="arch" type="xml">
             <xpath expr="//field[@name='product_id']" position="after">
-                <field name="yahoo_product_state_id" />
                 <field name="list_price" />
                 <field name="website_published" />
             </xpath>
         </field>
     </record>
-
     <record id="quant_search_view" model="ir.ui.view">
         <field name="name">stock.quant.search</field>
         <field name="model">stock.quant</field>
@@ -24,5 +21,4 @@
             </xpath>
         </field>
     </record>
-
 </odoo>
