+++ stock_quant_product_publish/__manifest__.py
@@ -1,0 +1,18 @@
+# Copyright 2017-2018 Quartile Limited
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
+{
+    "name": "Stock Quant Product Publish",
+    "version": "11.0.1.0.0",
+    "author": "Quartile Limited",
+    "website": "https://www.quartile.co",
+    "category": "Stock",
+    "license": "AGPL-3",
+    "depends": [
+        "stock",
+        "product_yahoo_auction_sst",
+    ],
+    "data": [
+        "wizards/stock_quant_product_publish_wizard.xml",
+    ],
+    "installable": True,
+}
