+++ stock_quant_product_publish/readme/DESCRIPTION.rst
@@ -1,0 +1,1 @@
+This module publishes products on the website and assigns the yahoo_product_state_id from stock.quant.
