+++ stock_quant_status_update/__manifest__.py
@@ -1,0 +1,19 @@
+# Copyright 2017-2018 Quartile Limited
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
+{
+    "name": "Stock Quant Status Update",
+    "version": "11.0.1.0.0",
+    "author": "Quartile Limited",
+    "website": "https://www.quartile.co",
+    "category": "Stock",
+    "license": "AGPL-3",
+    "depends": [
+        "stock",
+        "product_yahoo_auction_sst",
+    ],
+    "data": [
+        "views/res_config_settings_views.xml",
+        "wizards/stock_quant_status_update_wizard.xml",
+    ],
+    "installable": True,
+}
