+++ stock_quant_status_update/i18n/ja.po
@@ -1,0 +1,115 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+#	* stock_quant_status_update
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 11.0\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2023-10-02 12:34+0000\n"
+"PO-Revision-Date: 2023-10-02 12:34+0000\n"
+"Last-Translator: <>\n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: \n"
+
+#. module: stock_quant_status_update
+#: model:ir.ui.view,arch_db:stock_quant_status_update.stock_quant_status_update_wizard_form
+msgid "Cancel"
+msgstr "Cancel"
+
+#. module: stock_quant_status_update
+#: model:ir.model.fields,field_description:stock_quant_status_update.field_stock_quant_status_update_wizard_create_uid
+msgid "Created by"
+msgstr "Created by"
+
+#. module: stock_quant_status_update
+#: model:ir.model.fields,field_description:stock_quant_status_update.field_stock_quant_status_update_wizard_create_date
+msgid "Created on"
+msgstr "Created on"
+
+#. module: stock_quant_status_update
+#: model:ir.model.fields,field_description:stock_quant_status_update.field_stock_quant_status_update_wizard_display_name
+msgid "Display Name"
+msgstr "Display Name"
+
+#. module: stock_quant_status_update
+#: model:ir.model.fields,field_description:stock_quant_status_update.field_stock_quant_status_update_wizard_id
+msgid "ID"
+msgstr "ID"
+
+#. module: stock_quant_status_update
+#: model:ir.model.fields,field_description:stock_quant_status_update.field_stock_quant_status_update_wizard___last_update
+msgid "Last Modified on"
+msgstr "Last Modified on"
+
+#. module: stock_quant_status_update
+#: model:ir.model.fields,field_description:stock_quant_status_update.field_stock_quant_status_update_wizard_write_uid
+msgid "Last Updated by"
+msgstr "Last Updated by"
+
+#. module: stock_quant_status_update
+#: model:ir.model.fields,field_description:stock_quant_status_update.field_stock_quant_status_update_wizard_write_date
+msgid "Last Updated on"
+msgstr "Last Updated on"
+
+#. module: stock_quant_status_update
+#: model:ir.model,name:stock_quant_status_update.model_stock_quant
+msgid "Quants"
+msgstr "Quants"
+
+#. module: stock_quant_status_update
+#: model:ir.model.fields,field_description:stock_quant_status_update.field_res_config_settings_picking_product_state_id
+#: model:ir.ui.view,arch_db:stock_quant_status_update.res_config_settings_view_form
+msgid "State For Finished Picking Product"
+msgstr "State For Finished Picking Product"
+
+#. module: stock_quant_status_update
+#: model:ir.ui.view,arch_db:stock_quant_status_update.stock_quant_status_update_wizard_form
+msgid "Stock Quant Status Update"
+msgstr "Stock Quant Status Update"
+
+#. module: stock_quant_status_update
+#: model:ir.ui.view,arch_db:stock_quant_status_update.stock_quant_status_update_wizard_form
+msgid "This wizard will update the status of\n"
+"                    selected stock quants at once."
+msgstr "This wizard will update the status of\n"
+"                    selected stock quants at once."
+
+#. module: stock_quant_status_update
+#: model:ir.ui.view,arch_db:stock_quant_status_update.stock_quant_status_update_wizard_form
+msgid "Update"
+msgstr "Update"
+
+#. module: stock_quant_status_update
+#: model:ir.actions.act_window,name:stock_quant_status_update.stock_quant_status_update_wizard
+msgid "Update Stock Quant Status"
+msgstr "Update Stock Quant Status"
+
+#. module: stock_quant_status_update
+#: model:ir.model.fields,field_description:stock_quant_status_update.field_stock_quant_status_update_wizard_yahoo_product_state_id
+#: model:ir.ui.view,arch_db:stock_quant_status_update.res_config_settings_view_form
+msgid "Yahoo Product State"
+msgstr "Yahoo Product State"
+
+#. module: stock_quant_status_update
+#: model:ir.ui.view,arch_db:stock_quant_status_update.res_config_settings_view_form
+msgid "Yahoo Product State Settings"
+msgstr "Yahoo Product State Settings"
+
+#. module: stock_quant_status_update
+#: model:ir.ui.view,arch_db:stock_quant_status_update.stock_quant_status_update_wizard_form
+msgid "or"
+msgstr "or"
+
+#. module: stock_quant_status_update
+#: model:ir.model,name:stock_quant_status_update.model_res_config_settings
+msgid "res.config.settings"
+msgstr "res.config.settings"
+
+#. module: stock_quant_status_update
+#: model:ir.model,name:stock_quant_status_update.model_stock_quant_status_update_wizard
+msgid "stock.quant.status.update.wizard"
+msgstr "stock.quant.status.update.wizard"
