+++ stock_quant_status_update/models/stock_quant.py
@@ -1,31 +1,11 @@
 # Copyright 2018 Quartile Limited
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from odoo import fields, models
+from odoo import models
 
 
 class StockQuant(models.Model):
     _inherit = "stock.quant"
-
-    website_published = fields.Boolean(
-        related="product_id.product_tmpl_id.website_published",
-        string="Visible in Website",
-    )
-    yahoo_product_state_id = fields.Many2one(
-        related="product_id.product_tmpl_id.yahoo_product_state_id",
-        string="Yahoo Product State",
-    )
-    list_price = fields.Float(
-        related="product_id.product_tmpl_id.list_price", string="Sale Price",
-    )
-    product_id = fields.Many2one(auto_join=True,)
-    default_code = fields.Char(
-        related="product_id.default_code",
-        string="Barcode",
-        store=True,
-        readonly=True,
-        index=True,
-    )
 
     def update_stock_move_done_qty(self):
         stock_moves = self.env["stock.move"].search(
