+++ stock_quant_status_update/readme/DESCRIPTION.rst
@@ -1,0 +1,2 @@
+This module updates the yahoo_product_state_id in stock quant and adjusts the 'done' quantity in stock_move
+if the quant's yahoo_product_state_id matches the picking_product_state_id specified in res.config.settings.
