+++ stock_quant_status_update/wizards/__init__.py
@@ -1,2 +1,1 @@
-from . import stock_quant_publish_wizard
 from . import stock_quant_status_update_wizard
