+++ stock_quant_yahoo_product_state/__manifest__.py
@@ -1,0 +1,19 @@
+# Copyright 2017-2018 Quartile Limited
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
+{
+    "name": "Stock Quant Yahoo Product State",
+    "version": "11.0.1.0.0",
+    "author": "Quartile Limited",
+    "website": "https://www.quartile.co",
+    "category": "Stock",
+    "license": "AGPL-3",
+    "depends": [
+        "stock",
+        "stock_quant_list_view",
+        "product_yahoo_auction_sst",
+    ],
+    "data": [
+        "views/stock_quant_views.xml",
+    ],
+    "installable": True,
+}
