+++ stock_quant_yahoo_product_state/i18n/ja.po
@@ -1,0 +1,26 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+#	* stock_quant_yahoo_product_state
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 11.0\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2023-10-02 12:35+0000\n"
+"PO-Revision-Date: 2023-10-02 12:35+0000\n"
+"Last-Translator: <>\n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: \n"
+
+#. module: stock_quant_yahoo_product_state
+#: model:ir.model,name:stock_quant_yahoo_product_state.model_stock_quant
+msgid "Quants"
+msgstr "Quants"
+
+#. module: stock_quant_yahoo_product_state
+#: model:ir.model.fields,field_description:stock_quant_yahoo_product_state.field_stock_quant_yahoo_product_state_id
+msgid "Yahoo Product State"
+msgstr "Yahoo Product State"
