+++ stock_quant_yahoo_product_state/models/__init__.py
@@ -1,0 +1,1 @@
+from . import stock_quant
