+++ stock_quant_yahoo_product_state/readme/DESCRIPTION
@@ -1,0 +1,1 @@
+This module adds yahoo_product_state_id in stock.quant.
